+++ pyvkfft/test/test_fft.py
@@ -27,6 +27,7 @@ from pyvkfft.base import primes, radix_gen, radix_gen_n
 from pyvkfft.fft import fftn as vkfftn, ifftn as vkifftn, rfftn as vkrfftn, \
     irfftn as vkirfftn, dctn as vkdctn, idctn as vkidctn
 from pyvkfft.accuracy import test_accuracy, test_accuracy_kwargs, fftn, init_ctx, gpu_ctx_dic, has_dct_ref, has_scipy
+import pyvkfft.config
 
 try:
     import pycuda.gpuarray as cua
@@ -427,6 +428,9 @@ class TestFFT(unittest.TestCase):
         vtype = (np.complex64, np.complex128)
         if not has_cl_fp64:
             vtype = (np.complex64,)
+        # Disable warning
+        old_warning = pyvkfft.config.WARN_OPENCL_QUEUE_MISMATCH
+        pyvkfft.config.WARN_OPENCL_QUEUE_MISMATCH = False
         for dtype in vtype:
             with self.subTest(dtype=np.dtype(dtype)):
                 init_ctx("pyopencl", gpu_name=self.gpu, opencl_platform=self.opencl_platform, verbose=False)
@@ -451,27 +455,23 @@ class TestFFT(unittest.TestCase):
                 # test transforms with a supplied queue
                 for i in range(n_queues):
                     vapp[i].fft(vd[i], queue=queues[i])
-                for i in range(n_queues):
                     dn = fftn(np.roll(d, i * 7, axis=1))
                     self.assertTrue(np.allclose(dn, vd[i].get(), rtol=rtol, atol=abs(dn).max() * rtol))
 
                 for i in range(n_queues):
                     vapp[i].ifft(vd[i], queue=queues[i])
-                for i in range(n_queues):
                     dn = np.roll(d, i * 7, axis=1)
                     self.assertTrue(np.allclose(dn, vd[i].get(), rtol=rtol, atol=abs(dn).max() * rtol))
 
                 # test transforms with a supplied queue, different from the arrays'
                 for i in range(n_queues):
                     vapp[i].fft(vd[i], queue=queues[i + n_queues])
-                for i in range(n_queues):
                     dn = fftn(np.roll(d, i * 7, axis=1))
                     queues[i + n_queues].finish()
                     self.assertTrue(np.allclose(dn, vd[i].get(), rtol=rtol, atol=abs(dn).max() * rtol))
 
                 for i in range(n_queues):
                     vapp[i].ifft(vd[i], queue=queues[i + n_queues])
-                for i in range(n_queues):
                     dn = np.roll(d, i * 7, axis=1)
                     queues[i + n_queues].finish()
                     self.assertTrue(np.allclose(dn, vd[i].get(), rtol=rtol, atol=abs(dn).max() * rtol))
@@ -479,15 +479,61 @@ class TestFFT(unittest.TestCase):
                 # test transforms with the arrays queues
                 for i in range(n_queues):
                     vapp[i].fft(vd[i])
-                for i in range(n_queues):
                     dn = fftn(np.roll(d, i * 7, axis=1))
                     self.assertTrue(np.allclose(dn, vd[i].get(), rtol=rtol, atol=abs(dn).max() * rtol))
 
                 for i in range(n_queues):
                     vapp[i].ifft(vd[i])
-                for i in range(n_queues):
                     dn = np.roll(d, i * 7, axis=1)
                     self.assertTrue(np.allclose(dn, vd[i].get(), rtol=rtol, atol=abs(dn).max() * rtol))
+
+                # Test using the simple fft interface: inplace, same queue as array
+                for i in range(n_queues):
+                    vd[i] = vkfftn(vd[i], vd[i])
+                    dn = fftn(np.roll(d, i * 7, axis=1))
+                    self.assertTrue(np.allclose(dn, vd[i].get(), rtol=rtol, atol=abs(dn).max() * rtol))
+
+                    vd[i] = vkifftn(vd[i], vd[i])
+                    dn = np.roll(d, i * 7, axis=1)
+                    self.assertTrue(np.allclose(dn, vd[i].get(), rtol=rtol, atol=abs(dn).max() * rtol))
+
+                # Test using the simple fft interface: inplace, different queue for the transform
+                # A synchronisation is needed as get() will use the array's queue and not the transform's
+                for i in range(n_queues):
+                    vd[i] = vkfftn(vd[i], vd[i], cl_queue=queues[i + n_queues])
+                    dn = fftn(np.roll(d, i * 7, axis=1))
+                    queues[i + n_queues].finish()
+                    self.assertTrue(np.allclose(dn, vd[i].get(), rtol=rtol, atol=abs(dn).max() * rtol))
+
+                    vd[i] = vkifftn(vd[i], vd[i], cl_queue=queues[i + n_queues])
+                    dn = np.roll(d, i * 7, axis=1)
+                    queues[i + n_queues].finish()
+                    self.assertTrue(np.allclose(dn, vd[i].get(), rtol=rtol, atol=abs(dn).max() * rtol))
+
+                # Test using the simple fft interface: out-of-place, same queue as array
+                # The newly allocated array has the same queu as the source array, so no synchronisation issue
+                for i in range(n_queues):
+                    d1 = vkfftn(vd[i])
+                    dn = fftn(np.roll(d, i * 7, axis=1))
+                    self.assertTrue(np.allclose(dn, d1.get(), rtol=rtol, atol=abs(dn).max() * rtol))
+
+                    d2 = vkifftn(d1)
+                    dn = np.roll(d, i * 7, axis=1)
+                    self.assertTrue(np.allclose(dn, d2.get(), rtol=rtol, atol=abs(dn).max() * rtol))
+
+                # Test using the simple fft interface: out-of-place, different queue for the transform
+                # A synchronisation is needed as get() will use the array's queue and not the transform's
+                for i in range(n_queues):
+                    d1 = vkfftn(vd[i], cl_queue=queues[i + n_queues])
+                    dn = fftn(np.roll(d, i * 7, axis=1))
+                    queues[i + n_queues].finish()
+                    self.assertTrue(np.allclose(dn, d1.get(), rtol=rtol, atol=abs(dn).max() * rtol))
+
+                    d2 = vkifftn(d1, cl_queue=queues[i + n_queues])
+                    dn = np.roll(d, i * 7, axis=1)
+                    queues[i + n_queues].finish()
+                    self.assertTrue(np.allclose(dn, d2.get(), rtol=rtol, atol=abs(dn).max() * rtol))
+        pyvkfft.config.WARN_OPENCL_QUEUE_MISMATCH = old_warning
 
 
 # The class parameters are written in pyvkfft_test.main()